fix: Move `io_uring` into module under `vmm`
Moved `io_uring` into module under `vmm`.
