+++ src/vmm/src/io_uring/probe.rs
@@ -4,7 +4,7 @@
 use utils::fam::{FamStruct, FamStructWrapper};
 use utils::generate_fam_struct_impl;
 
-use crate::bindings::{io_uring_probe, io_uring_probe_op};
+use crate::io_uring::bindings::{io_uring_probe, io_uring_probe_op};
 
 // There is no max for the number of operations returned by probing. So we fallback to using the
 // number of values representable in a u8;
